Skaters Filters in URL
The Skaters page now builds the URL filters inside the controller with
only filters that have been changed from the default being placed in the
url (resulting in shorter URLS which is nice). Filters also include sort
parameter and sort order. Only the checkbox filters still need to be
implemented for the tables (could possibly do chart info in the URL as
well).